pyproject.toml
[tool.poetry]
name = "auctioneer-server"
version = "0.1.0"
description = ""
authors = ["Qurram Zaheer Syed <qurram.zaheer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-JWT-Extended = "^4.4.0"
python-dotenv = "^0.20.0"
pymongo = {extras = ["srv"], version = "^4.1.1"}
bcrypt = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
